+++ github3/repos.py
@@ -1138,33 +1138,6 @@ class Repository(GitHubCore):
             json = self._json(self._get(url), 200)
         return Milestone(json, self) if json else None
 
-    @requires_auth
-    def pubsubhubbub(self, mode, topic, callback, secret=''):
-        """Create/update a pubsubhubbub hook.
-
-        :param str mode: (required), accepted values: ('subscribe',
-            'unsubscribe')
-        :param str topic: (required), form:
-            https://github.com/:user/:repo/events/:event
-        :param str callback: (required), the URI that receives the updates
-        :param str secret: (optional), shared secret key that generates a
-            SHA1 HMAC of the payload content.
-        :returns: bool
-        """
-        from re import match
-        m = match('https://github\.com/\w+/[\w\._-]+/events/\w+', topic)
-        status = False
-        if mode and topic and callback and m:
-            data = [('hub.mode', mode), ('hub.topic', topic),
-                    ('hub.callback', callback)]
-            if secret:
-                data.append(('hub.secret', secret))
-            url = self._build_url('hub')
-            h = {'Content-Type': None}
-            status = self._boolean(self._post(url, data=data, headers=h), 204,
-                                   404)
-        return status
-
     def pull_request(self, number):
         """Get the pull request indicated by ``number``.